requirements.txt
psutil>=5.8.0,<6.0.0
pysmb==1.2.9.1
requests>=2.28.0
# 指定稳定的依赖版本以避免兼容性问题
# pysmb 1.2.9是已知在Docker环境中稳定工作的版本